+++ controlplane/test/federated-graph.test.ts
@@ -1,6 +1,6 @@
-import { createPromiseClient } from '@bufbuild/connect';
-import { fastifyConnectPlugin } from '@bufbuild/connect-fastify';
-import { createConnectTransport } from '@bufbuild/connect-node';
+import { createPromiseClient } from '@connectrpc/connect';
+import { fastifyConnectPlugin } from '@connectrpc/connect-fastify';
+import { createConnectTransport } from '@connectrpc/connect-node';
 import Fastify from 'fastify';
 import { pino } from 'pino';
 import { PlatformService } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_connect';